♻️ renderOrCopy と copyFile を render.go に移動
