+++ render.go
@@ -3,6 +3,8 @@ package main
 import (
 	"io"
 	"os"
+	"path/filepath"
+	"strings"
 	"text/template"
 
 	"github.com/Masterminds/sprig/v3"
@@ -36,4 +38,45 @@ func renderWithWriter(w io.Writer, templateFile string, input InputMap) error {
 		panic(err)
 	}
 	return nil
+}
+
+func renderOrCopy(srcFile, destPath string, templateExts []string, input InputMap) error {
+	isTemplate := false
+	for _, ext := range templateExts {
+		if strings.HasSuffix(destPath, ext) {
+			isTemplate = true
+			destPath = strings.TrimSuffix(destPath, ext)
+			break
+		}
+	}
+
+	// Write the rendered content to the destination directory
+	if err := os.MkdirAll(filepath.Dir(destPath), 0o755); err != nil {
+		return err
+	}
+
+	if isTemplate {
+		return renderToFile(srcFile, input, destPath)
+	} else {
+		// Copy non-template files directly
+		return copyFile(srcFile, destPath)
+	}
+}
+
+// copyFile copies a file from src to dst
+func copyFile(src, dst string) error {
+	sourceFile, err := os.Open(src)
+	if err != nil {
+		return err
+	}
+	defer sourceFile.Close()
+
+	destinationFile, err := os.Create(dst)
+	if err != nil {
+		return err
+	}
+	defer destinationFile.Close()
+
+	_, err = io.Copy(destinationFile, sourceFile)
+	return err
 }
